+++ lib/id_doc_kit.dart
@@ -1,4 +1,4 @@
 export 'src/models/id_document_type.dart';
 export 'src/models/id_document_result.dart';
 export 'src/validators/id_validator.dart';
-export 'src/widgets/id_text_field.dart';
+export 'src/widgets/id_fields.dart';