+++ lib/src/models/id_document_type.dart
@@ -1,5 +1,1 @@
-enum IdDocumentType {
-  aadhaar,
-  pan,
-  drivingLicense,
-}
+enum IdDocumentType { aadhaar, pan, drivingLicense, gstin }