+++ lib/src/validators/id_validator.dart
@@ -3,6 +3,7 @@ import '../models/id_document_result.dart';
 import 'aadhaar_validator.dart';
 import 'pan_validator.dart';
 import 'dl_validator.dart';
+import 'gstin_validator.dart';
 
 class IdValidator {
   IdValidator._internal();
@@ -12,6 +13,7 @@ class IdValidator {
   final _aadhaar = AadhaarValidator();
   final _pan = PanValidator();
   final _dl = DrivingLicenseValidator();
+  final _gstin = GstinValidator();
 
   IdDocumentResult validate({
     required IdDocumentType type,
@@ -24,6 +26,8 @@ class IdValidator {
         return _pan.validate(value);
       case IdDocumentType.drivingLicense:
         return _dl.validate(value);
+      case IdDocumentType.gstin:
+        return _gstin.validate(value);
     }
   }
 
@@ -35,6 +39,10 @@ class IdValidator {
       return validate(type: IdDocumentType.aadhaar, value: value);
     } else if (RegExp(r'^[A-Z]{5}[0-9]{4}[A-Z]$').hasMatch(cleaned)) {
       return validate(type: IdDocumentType.pan, value: value);
+    } else if (RegExp(
+      r'^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z][0-9A-Z]$',
+    ).hasMatch(cleaned)) {
+      return validate(type: IdDocumentType.gstin, value: value);
     } else {
       // Try DL as a last attempt
       final dlResult = validate(